requirements.txt
pyTelegramBotAPI
requests
schedule
psycopg2-binary
